db-explorer/pyproject.toml
[project]
name = "db-explorer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "psycopg2-binary>=2.9.11",
    "pysqlite3>=0.6.0",
]
